+++ README.ru.md
@@ -342,6 +342,19 @@ sudo journalctl -fu ssp.service
 }
 ```
 
+
+## Поддержка реализаций LDAP
+Портал протестирован со следующими реализации каталогов:
+ - ActiveDirectory
+ - Samba4
+ - FreeIPA
+### Особенности при работе с FreeIPA
+Для корректного подключения к FreeIPA необходимо в файле конфигурации раскомментировать и заполнить настройку `LdapBaseDn`. Пример:
+```xml
+<LdapBaseDn>cn=users,cn=accounts,dc=domain,dc=local</LdapBaseDn>
+```
+
+
 ## Сценарии использования
 
 Портал используется для самостоятельной регистрации второго фактора аутентификации пользователями внутри корпоративной сети, а также выполняет роль единой точки входа для приложений, работающих по технологии Single Sign-On.
