+++ src/MultiFactor.SelfService.Linux.Portal/Extensions/ServicesConfiguring.cs
@@ -2,12 +2,13 @@
 using MultiFactor.SelfService.Linux.Portal.Abstractions.Ldap;
 using MultiFactor.SelfService.Linux.Portal.Authentication;
 using MultiFactor.SelfService.Linux.Portal.Core.Http;
-using MultiFactor.SelfService.Linux.Portal.Integrations.ActiveDirectory.CredentialVerification;
 using MultiFactor.SelfService.Linux.Portal.Integrations.ActiveDirectory.ExchangeActiveSync;
-using MultiFactor.SelfService.Linux.Portal.Integrations.ActiveDirectory.PasswordChanging;
 using MultiFactor.SelfService.Linux.Portal.Integrations.Google;
 using MultiFactor.SelfService.Linux.Portal.Integrations.Google.ReCaptcha;
 using MultiFactor.SelfService.Linux.Portal.Integrations.Ldap;
+using MultiFactor.SelfService.Linux.Portal.Integrations.Ldap.Connection;
+using MultiFactor.SelfService.Linux.Portal.Integrations.Ldap.CredentialVerification;
+using MultiFactor.SelfService.Linux.Portal.Integrations.Ldap.PasswordChanging;
 using MultiFactor.SelfService.Linux.Portal.Integrations.MultiFactorApi;
 using MultiFactor.SelfService.Linux.Portal.Settings;
 using MultiFactor.SelfService.Linux.Portal.Stories.AddYandexAuthStory;
@@ -42,14 +43,18 @@ namespace MultiFactor.SelfService.Linux.Portal.Extensions
                 .AddSingleton<JsonDataSerializer>()
                 .AddSingleton<JsonPayloadLogger>()
                 .AddSingleton<DeviceAccessStateNameLocalizer>()
-                .AddSingleton<ActiveDirectoryCredentialVerifier>()
-                .AddSingleton<ActiveDirectoryPasswordChanger>()
+                .AddSingleton<CredentialVerifier>()
+                .AddSingleton<PasswordChanger>()
                 .AddSingleton<HttpClientTokenProvider>()
                 .AddSingleton<ExchangeActiveSyncDevicesSearcher>()
                 .AddSingleton<ExchangeActiveSyncDeviceStateChanger>()
                 .AddSingleton<LdapConnectionAdapterFactory>()
+
                 .AddSingleton<LdapBindDnFormatterFactory>()
                 .AddSingleton(services => services.GetRequiredService<LdapBindDnFormatterFactory>().CreateFormatter())
+
+                .AddSingleton<PasswordAttributeReplacerFactory>()
+                .AddSingleton(services => services.GetRequiredService<PasswordAttributeReplacerFactory>().CreateReplacer())
 
                 .AddTransient<HttpMessageInterceptor>()
                 .AddTransient<ICaptchaVerifier, GoogleReCaptchaVerifier>()
