+++ src/MultiFactor.SelfService.Linux.Portal/Integrations/ActiveDirectory/ExchangeActiveSync/ExchangeActiveSyncDevicesSearcher.cs
@@ -2,6 +2,7 @@
 using MultiFactor.SelfService.Linux.Portal.Abstractions.Ldap;
 using MultiFactor.SelfService.Linux.Portal.Integrations.ActiveDirectory.ExchangeActiveSync.Models;
 using MultiFactor.SelfService.Linux.Portal.Integrations.Ldap;
+using MultiFactor.SelfService.Linux.Portal.Integrations.Ldap.Connection;
 using MultiFactor.SelfService.Linux.Portal.Settings;
 using System.Globalization;
 using static LdapForNet.Native.Native;
