--- src/MultiFactor.SelfService.Linux.Portal/Integrations/ActiveDirectory/PasswordChanging/PasswordChangingResult.cs
@@ -1,4 +1,0 @@
-﻿namespace MultiFactor.SelfService.Linux.Portal.Integrations.ActiveDirectory.PasswordChanging
-{
-    public record PasswordChangingResult(bool Success, string ErrorReason);
-}
