+++ src/MultiFactor.SelfService.Linux.Portal/Integrations/Ldap/Connection/LdapConnectionAdapterConfig.cs
@@ -1,0 +1,10 @@
+﻿using MultiFactor.SelfService.Linux.Portal.Abstractions.Ldap;
+
+namespace MultiFactor.SelfService.Linux.Portal.Integrations.Ldap.Connection
+{
+    public class LdapConnectionAdapterConfig
+    {
+        public ILdapBindDnFormatter Formatter { get; set; } = new DefaultLdapBindDnFormatter();
+        public ILogger? Logger { get; set; }
+    }
+}
