+++ src/MultiFactor.SelfService.Linux.Portal/Integrations/Ldap/PasswordChanging/PasswordChangingResult.cs
@@ -1,0 +1,4 @@
+﻿namespace MultiFactor.SelfService.Linux.Portal.Integrations.Ldap.PasswordChanging
+{
+    public record PasswordChangingResult(bool Success, string ErrorReason);
+}
